+++ addon/components/bootstrap/form/-controls.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+import layout from '../../../templates/components/bootstrap/form/-controls';
+
+export default Ember.Component.extend({
+  layout,
+  tagName: '',
+});