+++ addon/components/bootstrap/simple-form.js
@@ -1,0 +1,17 @@
+import Ember from 'ember';
+import layout from '../../templates/components/bootstrap/simple-form';
+import { PropTypes } from 'ember-prop-types';
+import { BuilderForPropTypes, BuilderForPropDefaults } from 'ember-bootstrap-controls/utils/prop-definition-tools';
+
+export const propDefinitions = {
+};
+
+export default Ember.Component.extend({
+  layout,
+  tagName: 'form',
+  propTypes: BuilderForPropTypes(propDefinitions),
+
+  getDefaultProps() {
+    return BuilderForPropDefaults(propDefinitions)
+  },
+});