+++ addon/components/freestyle/bootstrap/form/-controls.js
@@ -1,0 +1,14 @@
+import Ember from 'ember';
+import layout from '../../../templates/components/freestyle/bootstrap/form/-controls';
+import { BuilderForPropDefaults } from 'ember-bootstrap-controls/utils/prop-definition-tools';
+import { propDefinitions } from '../../bootstrap/form/-controls';
+
+export default Ember.Component.extend({
+  layout,
+  propDefinitions,
+  data: Object.assign(BuilderForPropDefaults(propDefinitions), {
+    value: '',
+  }),
+
+  basicValue: '',
+});