+++ tests/integration/components/bootstrap/simple-form-test.js
@@ -1,0 +1,31 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+import a11yAudit from 'ember-a11y-testing/test-support/audit';
+
+moduleForComponent('bootstrap/simple-form', 'Integration | Component | bootstrap button', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+  // Example if your component had an `action` that it required:
+  // this.set('noop', () => {});
+  // this.render(hbs`{{bootstrap/simple-form action=noop}}`);
+  this.render(hbs`{{bootstrap/simple-form}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#bootstrap/simple-form}}
+      template block text
+    {{/bootstrap/simple-form}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+
+  return a11yAudit(this.$()).then(() => {
+    assert.ok(true, 'no a11y errors found!');
+  });
+});